Use uint16 instead of int32. Small speed and memory footprint improvement.

+++ src/entropy/AdaptiveProbMap.hpp
@@ -20,12 +20,12 @@ limitations under the License.
 
 // APM maps a probability and a context into a new probability
 // that the next bit will be 1. After each guess, it updates
-// its state to improve future guesses. 
+// its state to improve future guesses.
 
-namespace kanzi 
+namespace kanzi
 {
    template <int RATE>
-   class LinearAdaptiveProbMap 
+   class LinearAdaptiveProbMap
    {
    public:
        LinearAdaptiveProbMap<RATE>(int n);
@@ -36,21 +36,21 @@ namespace kanzi
 
    private:
        int _index; // last p, context
-       int32* _data; // [NbCtx][33]:  p, context -> p
+       uint16* _data; // [NbCtx][33]:  p, context -> p
    };
 
    template <int RATE>
    inline LinearAdaptiveProbMap<RATE>::LinearAdaptiveProbMap(int n)
    {
-       _data = new int32[65 * n];
+       _data = new uint16[65 * n];
        _index = 0;
 
        for (int j = 0; j <= 64; j++) {
-          _data[j] = int32(j << 6) << 4;
+          _data[j] = uint16(j << 6) << 4;
        }
 
        for (int i = 1; i < n; i++) {
-           memcpy(&_data[i*65], &_data[0], 65*sizeof(int32));
+           memcpy(&_data[i*65], &_data[0], 65*sizeof(uint16));
        }
    }
 
@@ -59,21 +59,21 @@ namespace kanzi
    inline int LinearAdaptiveProbMap<RATE>::get(int bit, int pr, int ctx)
    {
        // Update probability based on error and learning rate
-       const int32 g = int32((bit << 16) + (bit << RATE) - (bit << 1));
-       _data[_index] += ((g - _data[_index]) >> RATE);
-       _data[_index + 1] += ((g - _data[_index + 1]) >> RATE);
+       const int g = (bit << 16) + (bit << RATE) - (bit << 1);
+       _data[_index] += uint16((g - int(_data[_index])) >> RATE);
+       _data[_index + 1] += uint16((g - int(_data[_index + 1])) >> RATE);
 
        // Find index: 65*ctx + quantized prediction in [0..64]
        _index = (pr >> 6) + (ctx << 6) + ctx;
 
        // Return interpolated probabibility
-       const int32 w = int32(pr & 127);
+       const uint16 w = uint16(pr & 127);
        return int(_data[_index] * (128 - w) + _data[_index + 1] * w) >> 11;
    }
 
 
    template <int RATE>
-   class LogisticAdaptiveProbMap 
+   class LogisticAdaptiveProbMap
    {
    public:
        LogisticAdaptiveProbMap<RATE>(int n);
@@ -84,21 +84,21 @@ namespace kanzi
 
    private:
        int _index; // last p, context
-       int32* _data; // [NbCtx][33]:  p, context -> p
+       uint16* _data; // [NbCtx][33]:  p, context -> p
    };
 
    template <int RATE>
    inline LogisticAdaptiveProbMap<RATE>::LogisticAdaptiveProbMap(int n)
    {
-       _data = new int32[33 * n];
+       _data = new uint16[33 * n];
        _index = 0;
 
        for (int j = 0; j <= 32; j++) {
-          _data[j] = int32(Global::squash((j - 16) << 7)) << 4;
+          _data[j] = uint16(Global::squash((j - 16) << 7)) << 4;
        }
 
        for (int i = 1; i < n; i++) {
-           memcpy(&_data[i*33], &_data[0], 33*sizeof(int32));
+           memcpy(&_data[i*33], &_data[0], 33*sizeof(uint16));
        }
    }
 
@@ -107,16 +107,16 @@ namespace kanzi
    inline int LogisticAdaptiveProbMap<RATE>::get(int bit, int pr, int ctx)
    {
        // Update probability based on error and learning rate
-       const int32 g = int32((bit << 16) + (bit << RATE) - (bit << 1));
-       _data[_index] += ((g - _data[_index]) >> RATE);
-       _data[_index + 1] += ((g - _data[_index + 1]) >> RATE);
+       const int g = (bit << 16) + (bit << RATE) - (bit << 1);
+       _data[_index] += uint16((g - int(_data[_index])) >> RATE);
+       _data[_index + 1] += uint16((g - int(_data[_index + 1])) >> RATE);
        pr = Global::STRETCH[pr];
 
        // Find index: 33*ctx + quantized prediction in [0..32]
        _index = ((pr + 2048) >> 7) + (ctx << 5) + ctx;
 
        // Return interpolated probabibility
-       const int32 w = int32(pr & 127);
+       const uint16 w = uint16(pr & 127);
        return int(_data[_index] * (128 - w) + _data[_index + 1] * w) >> 11;
    }
 
@@ -132,22 +132,22 @@ namespace kanzi
 	   int get(int bit, int pr, int ctx);
 
    private:
-	   int32* _p; // last p
-	   int32* _data; // [NbCtx][33]:  p, context -> p
+	   uint16* _p; // last p
+	   uint16* _data; // [NbCtx][33]:  p, context -> p
    };
 
    template <int RATE>
    inline FastLogisticAdaptiveProbMap<RATE>::FastLogisticAdaptiveProbMap(int n)
    {
-	   _data = new int32[33 * n];
+	   _data = new uint16[33 * n];
 	   _p = &_data[0];
 
 	   for (int j = 0; j <= 32; j++) {
-		   _data[j] = int32(Global::squash((j - 16) << 7)) << 4;
+		   _data[j] = uint16(Global::squash((j - 16) << 7)) << 4;
 	   }
 
 	   for (int i = 1; i < n; i++) {
-		   memcpy(&_data[i * 33], &_data[0], 33 * sizeof(int32));
+		   memcpy(&_data[i * 33], &_data[0], 33 * sizeof(uint16));
 	   }
    }
 
@@ -156,8 +156,8 @@ namespace kanzi
    inline int FastLogisticAdaptiveProbMap<RATE>::get(int bit, int pr, int ctx)
    {
 	   // Update probability based on error and learning rate
-	   const int32 g = int32((bit << 16) + (bit << RATE) - (bit << 1));
-	   *_p += ((g - *_p) >> RATE);
+	   const int g = (bit << 16) + (bit << RATE) - (bit << 1);
+	   *_p += ((g - int(*_p)) >> RATE);
 
 	   // Find index: 33*ctx + quantized prediction in [0..32]
 	   _p = &_data[((Global::STRETCH[pr] + 2048) >> 7) + (ctx << 5) + ctx];
